+++ GrowthBook/src/commonMain/kotlin/com/sdk/growthbook/model/GBContext.kt
@@ -3,6 +3,7 @@
 package com.sdk.growthbook.model
 
 import com.sdk.growthbook.GBTrackingCallback
+import com.sdk.growthbook.plugin.tracking.GrowthBookPlugin
 import com.sdk.growthbook.utils.GBFeatures
 import com.sdk.growthbook.utils.GBStickyAssignmentsDocument
 import com.sdk.growthbook.utils.GBStickyAttributeKey
@@ -108,7 +109,17 @@ class GBContext(
      */
     val enableLogging: Boolean = false,
 
+    /**
+     * Saved Groups let you target the same group of users across multiple features and experiments.
+     */
     savedGroups: Map<String, GBValue>? = null,
+
+    /**
+     * Plugins registered with the GrowthBook. See
+     * [GrowthBookPlugin] and
+     * [com.sdk.growthbook.plugin.tracking.GrowthBookTrackingPlugin].
+     */
+    var plugins: List<GrowthBookPlugin>? = null
 ) {
 
     // Single source of truth for all cross-thread-shared evaluation inputs.