+++ GrowthBook/src/jvmTest/kotlin/com/sdk/growthbook/tests/GBExperimentRunTests.kt
@@ -91,7 +91,8 @@ class GBExperimentRunTests {
                         attributes = gbContext.attributes,
                         stickyBucketAssignmentDocs = gbContext.stickyBucketAssignmentDocs,
                     ),
-                    stackContext = StackContext(null, mutableSetOf())
+                    stackContext = StackContext(null, mutableSetOf()),
+                    pluginRegistry = null
                 )
 
                 val evaluator = GBExperimentEvaluator(testScopeEvaluationContext)
@@ -161,7 +162,8 @@ class GBExperimentRunTests {
                     qaMode = testContext.qaMode,
                     stickyBucketAssignmentDocs = null,
                 ),
-                stackContext = StackContext(null, mutableSetOf())
+                stackContext = StackContext(null, mutableSetOf()),
+                pluginRegistry = null
 
             )